Fixing the bugs. Need to save midway

+++ s21_string+/src/s21_sprintf/s21_itoa.h
@@ -4,7 +4,7 @@
 #include <stdio.h>
 #include <math.h>
 
-char* s21_itoa(int number, char *buffer, int base, int *number_of_digits_Out);
-char* s21_itoa_unsigned(unsigned int number, char *buffer, int base, int *number_of_digits_Out);
+char* s21_itoa(long long int number, char *buffer, int base, int *number_of_digits_Out);
+char* s21_itoa_unsigned(unsigned long long int number, char *buffer, int base, int *number_of_digits_Out);
 
 #endif
